In game I have removed player1/2 from the while body to reduce the recal of the constructor and i will just reset to the initial state. To reset them I added the ResetVector methode in Player class and also added the chooseCard methode.

+++ Eter/Eter/SimpleCard.cpp
@@ -43,6 +43,6 @@ void SimpleCard::setColor(std::string_view color)
 
 std::ostream& operator<<(std::ostream& os, const SimpleCard& card)
 {
-	os << card.m_value << card.m_color;
+	os << static_cast<int>(card.m_value) << card.m_color;
 	return os;
 }
